+++ app/settings.py
@@ -33,18 +33,6 @@ PLACEHOLDER = "string"  # Swagger UI placeholder value
 # Fonts
 
 DEFAULT_FONT = "thick"
-FONT_PATHS = {
-    # Names
-    "impact": ROOT / "fonts" / "Impact.ttf",
-    "kalam": ROOT / "fonts" / "Kalam-Regular.ttf",
-    "notosans": ROOT / "fonts" / "NotoSans-Bold.ttf",
-    "titilliumweb": ROOT / "fonts" / "TitilliumWeb-Black.ttf",
-    # Aliases
-    "thick": ROOT / "fonts" / "TitilliumWeb-Black.ttf",
-    "thin": ROOT / "fonts" / "TitilliumWeb-SemiBold.ttf",
-    "tiny": ROOT / "fonts" / "Segoe UI Bold.ttf",
-    "comic": ROOT / "fonts" / "Kalam-Regular.ttf",
-}
 
 MINIMUM_FONT_SIZE = 7